Refactor to use libgfi
