--- data/gfiLabelAliases.json
@@ -1,10 +1,0 @@
-[
-  "\"good first issue\"",
-  "good-first-issue",
-  "\"Good first issue\"",
-  "\"good first issue :octocat:\"",
-  "\":wave: Good First Issue\"",
-  "\"Good for new contributors\"",
-  "\"Good First Contribution\"",
-  "Suggested"
-]
