--- examples/README.md
@@ -1,10 +1,0 @@
-# Good First Issue: Examples
-
-Good First Issue can be consumed as a Node.js module. This directory includes examples of how you can use it to programatically surface Good First Issues.
-
-## Examples
-
-- [node.js](./node.js): An example of how you can consume the module using async/await and stitch together data from Good First Issue to create useful output without other external dependencies.
-- [github.js](./github.js): An example that's almost identical to the `node.js` example, except it utilizes the org/repo search functionality rather than using a project identifier.
-
-If you'd like to see another example, please feel free to sumbit a PR – we'd love to have your contribution! 🤗
